test(fuzz): add native-smoke harness path for macOS (closes #920)
The pinned nightly's ASan runtime deadlocks in dyld before main on
macOS 26.5, so `cargo fuzz run` never executes an input on a Mac. Extract
each bridge fuzz target's body into the botho-fuzz library so the
libFuzzer target and a new native-smoke driver share one source of truth,
then run the shared bodies over the corpus + randomized inputs with no
libFuzzer/sanitizer.

- Move the fuzz_bridge_attestation / fuzz_bridge_solana_parse /
  fuzz_federation_status_line harness bodies (Arbitrary types + all
  invariant assertions) into fuzz/src/*.rs; the fuzz_targets/*.rs binaries
  become one-line wrappers calling the shared run() fn. libFuzzer and the
  native path decode bytes identically (arbitrary_take_rest), so a corpus
  seed drives the same code path in both worlds — no harness drift.
- Add a `native-smoke` feature with a #[cfg(test)] suite (one test per
  target) and a `native_smoke` bin (~100k dev / configurable CI budget)
  that replay corpus seeds + randomized inputs on a stock toolchain.
- Document the macOS 26.5 ASan/dyld deadlock (with the sample stack) and
  the native-smoke commands in fuzz/README.md; note coverage-guided runs
  happen in ubuntu CI.
- Add a toolchain-watch note in fuzz/Cargo.toml listing the two
  constraints a nightly bump must satisfy, linking #920.
- Piggyback a cheap native-smoke job on the PR fuzz-build guard in
  fuzz.yml so it runs on ubuntu too.
